Don’t need to retry on outer timeout with fixed retransmit timeout
Now that the UDP RPC client is configured to retransmit after a
sensible delay, we will never hit an overall RPC timeout, so we don't
need the ugly code we had in there to deal with that.

But all the clients need to be configured this way. (In the real final
implementation we will keep a single client around for each service,
rather than creating them on each call like this.)