Fix JavaDoc builds under JDK 11 again.
It seems there are still fluctuations in the JDK and maven plugins,
but hopefully this approach is going to be stable, and gets the search
links to actually work when JavaDoc is built under JDK 11, w00t!

The new minimum build for JDKs is 9. This change also fixes the build
to be sure it compiles against compatible boot path classes for the
target release of Java 1.6 regardless of the build version.