Fix parsing of phrases if mood unknown.
Reviewing the diffs in
https://github.com/EvanPurkhiser/prolink-connect/commit/6db259ff26793d872398e94b47f5369a21825acb
I realized I was using `unknown_tag` in a context that actually would
not work, and would lead to a failed parse of the phrase entries if we
ever encounter a file with a new mood. This change at least fails
safely, in the sense that we just interpret them as mid-mood phrases
until we can gather more information about the new mood.